requirements.txt
python-docx>=1.1.0
bayoo-docx>=0.0.1
docx2python>=3.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
openai>=1.54.0
anthropic>=0.39.0
lxml>=5.0.0
python-dotenv>=1.0.0
pydantic==2.10.5
pydantic-core==2.27.2
typing-extensions>=4.8.0
flask>=3.0.0
werkzeug>=3.0.0
